% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetDocumentation.R
\docType{data}
\name{abcdPlotGpData}
\alias{abcdPlotGpData}
\title{abcdPlotGpData}
\format{A data set with columns
\describe{
  \item{PlotID}{A unique identifer for the plot. The first four letters indicate the park.}
  \item{dayN}{Day of Year}
  \item{yearV}{Year}
  \item{NdviMed}{NDVI level. Median NDVI of all sampled cells}
  \item{CellCntScrn}{Number of cells possible for the site.}
  \item{CellCnt}{Number of cells actually used}
  }}
\usage{
abcdPlotGpData
}
\description{
A data set with the following attributes
}
\keyword{datasets}

